% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msm_fit.R
\name{msm_fit}
\alias{msm_fit}
\title{Fit marginal structural model for two mixtures}
\usage{
msm_fit(
  f,
  data,
  mix1,
  mix2,
  interaction = TRUE,
  q = 4,
  id = NULL,
  MCsize = nrow(data)
)
}
\arguments{
\item{f}{Model formula}

\item{data}{Data frame}

\item{mix1}{Names of mixture 1 variables}

\item{mix2}{Names of mixture 2 variables}

\item{interaction}{Include interaction term}

\item{q}{Number of quantiles}

\item{id}{Cluster id}

\item{MCsize}{Monte Carlo sample size}
}
\value{
Estimated MSM coefficients
}
\description{
Fit marginal structural model for two mixtures
}
