% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgcomp.glm.multi.boot.R
\name{qgcomp.glm.multi.boot}
\alias{qgcomp.glm.multi.boot}
\title{Fit quantile g-computation model for two mixtures}
\usage{
qgcomp.glm.multi.boot(
  f,
  data,
  mix1,
  mix2,
  interaction = TRUE,
  q = 4,
  B = 200,
  id = NULL,
  MCsize = nrow(data)
)
}
\arguments{
\item{f}{Model formula}

\item{data}{Data frame}

\item{mix1}{Names of mixture 1 variables}

\item{mix2}{Names of mixture 2 variables}

\item{interaction}{Include interaction term}

\item{q}{Number of quantiles}

\item{B}{Number of bootstrap iterations}

\item{id}{Cluster id}

\item{MCsize}{Monte Carlo sample size}
}
\value{
Fitted qgcomp model
}
\description{
Fit quantile g-computation model for two mixtures
}
