% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantize_mixtures.R
\name{quantize_mixtures}
\alias{quantize_mixtures}
\title{Quantize mixture variables}
\usage{
quantize_mixtures(data, mix1, mix2, q = 4)
}
\arguments{
\item{data}{Data frame}

\item{mix1}{Names of mixture 1 variables}

\item{mix2}{Names of mixture 2 variables}

\item{q}{Number of quantiles}
}
\value{
Data frame with quantized variables
}
\description{
Converts dataset with continuous mixture components into quantized version.
}
