% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgcompmulti-diagnostics.R
\name{support}
\alias{support}
\title{Extract intervention support diagnostics from a qgcompmulti fit}
\usage{
support(object, ...)
}
\arguments{
\item{object}{A fitted \code{"qgcompmulti"} object.}

\item{...}{Unused.}
}
\value{
An object of class \code{"qgcompmulti_support_diagnostic"}.
}
\description{
Extract intervention support diagnostics from a qgcompmulti fit
}
